refactor(ui): Refactor dropdown to be more generic
- Combine sidebarDropdown with BaseDropdown
- Add anchor id for input
- Remove unnecessary div tags
