--- client/components/Dropdown/Base/BaseDropdown.tsx
@@ -1,82 +1,0 @@
-import React, { FC, useEffect, useState } from 'react';
-import { fromEvent, Subscription, merge } from 'rxjs';
-import Portal from '@/components/Portal/Portal';
-import styles from './BaseDropdown.scss';
-import { DropdownMenuItem } from './models/DropdownMenuItem';
-import { BaseDropdownProps } from './models/BaseDropdown';
-
-const BaseDropdown: FC<BaseDropdownProps> = ({
-  selected,
-  menuItems,
-  rect,
-  title,
-  rowClick,
-  close,
-}) => {
-  const [dropdownStyle, setDropdownStyle] = useState([styles.dropdown]);
-
-  // Event handler for closing dropdown
-  const handleCloseEvent = (event: Event) => {
-    const modal = document.getElementById('dropdown');
-    if (event.type === 'resize' || !modal.contains(event.target as Node)) {
-      close();
-    }
-  };
-
-  // Event handler for row clicks
-  const handleRowClick = (menuItem: DropdownMenuItem) => {
-    rowClick(menuItem);
-    close();
-  };
-
-  // Event listener for outside click dropdown click
-  useEffect(() => {
-    // Delayed style for animation
-    setDropdownStyle(dropdownStyle.concat(styles.dropdownOpen));
-
-    // Close event on clicks and resize
-    const subscription: Subscription = merge(
-      fromEvent(document, 'click'),
-      fromEvent(window, 'resize'),
-    ).subscribe(handleCloseEvent);
-
-    return () => subscription.unsubscribe();
-  }, []);
-
-  // Button class based on row
-  const buttonClass = (value: string) => {
-    const className = [styles.item];
-    if (selected === value) {
-      className.push(styles.itemSelected);
-    }
-
-    return className.join(' ');
-  };
-
-  return (
-    <Portal rect={rect} target="dropdown">
-      <div className={dropdownStyle.join(' ')}>
-        {title && (
-          <div className={styles.header}>
-            <h3>{title}</h3>
-          </div>
-        )}
-        <div className={styles.content}>
-          {menuItems.map(menuItem => (
-            <button
-              key={menuItem.label}
-              type="button"
-              onClick={() => handleRowClick(menuItem)}
-              className={buttonClass(menuItem.label)}
-              style={menuItem.style}
-            >
-              {menuItem.label}
-            </button>
-          ))}
-        </div>
-      </div>
-    </Portal>
-  );
-};
-
-export default BaseDropdown;
