--- client/components/Dropdown/Sidebar/models/SidebarDropdown.ts
@@ -1,8 +1,0 @@
-import { DropdownMenuItem } from '../../Base/models/DropdownMenuItem';
-
-export interface SidebarDropdownProps {
-  selected: string;
-  menuItems: DropdownMenuItem[];
-  close(): void;
-  rowClick(item: DropdownMenuItem): void;
-}
