+++ client/components/Dropdown/models/Offset.ts
@@ -1,0 +1,4 @@
+export interface Offset {
+  x?: number;
+  y?: number;
+}
